Add support for gpDAQ A2D calibrations.
This really meant detecting the gpDAQ cals in the DB, and exporting to nidas A2D cal file.

The main need for this is two fold, the gpDAQ only has 4 channels and Josh's requirement to support 3rd order calibrations for the gp and temp DAQs.

 - Added new sensor_type 'analog_gpDAQ'.  This allows differentiating between NCAR A2D card and this.
 - 'cal_type' still remains with three options, 'bath', 'analog', and 'instrument'.
 - Output 4 channels and 4 coefficients per channel (vs 8 and 2).

One requirement to group all the channels was a three day limit.  JAM had already done an over the weekend cal that exceeded three days, and the software didn't detect it as a group.  Increase to five days.